Create case a user has an API
- Create a new component Date range picker
- Separate API Analytic page to header, body, date range picker templates
- Display API Ananlytic header & body if a user has an API is connected to current proxy
