+++ apinf_packages/core/lib/i18n/en.i18n.json
@@ -30,12 +30,9 @@
   "addOrganizationManagerForm_description": "An organization manager has the same rights as the original API manager to edit, delete the API after an API has been connected to an organization.",
   "addOrganizationManagerForm_button": "Add",
   "organizationManagerForm_managerAlreadyExist_errorText": "User is already a manager of this organization.",
-  "apiAnalyticPage_optionText_yesterday": "Yesterday",
-  "apiAnalyticPage_optionText_7days": "Last 7 Days",
-  "apiAnalyticPage_optionText_30days": "Last 30 Days",
-  "apiAnalyticPage_labelText_proxyName": "Proxy: __name__",
-  "apiAnalyticPage_warningText_noData": "There are no API usage statistics available.",
-  "apiAnalyticPage_text_elasticsearchError": "Something went wrong. Please try again",
+  "apiAnalyticPageBody_warningText_noData": "There are no API usage statistics available.",
+  "apiAnalyticPageBody_text_elasticsearchError": "Something went wrong. Please try again",
+  "apiAnalyticPageHeader_labelText_proxyName": "Proxy: __name__",
   "apiAnalyticView_text_requestsCount": "__count__ request",
   "apiAnalyticView_text_requestsCount_plural": "__count__ requests",
   "apiAnalyticView_text_responseTime": "__time__ ms",
@@ -258,14 +255,14 @@
   "dashboardToolbar_optionText_calls": "Highest calls",
   "dashboardToolbar_optionText_users": "Most users",
   "dashboardToolbar_optionText_time": "Highest response time",
-  "dashboardToolbar_optionText_yesterday": "Yesterday",
-  "dashboardToolbar_optionText_7days": "Last 7 Days",
-  "dashboardToolbar_optionText_30days": "Last 30 Days",
   "dashboardPage_buttonText_addApi": "Add an API",
   "dashboardPage_title_dashboard": "Dashboard",
   "dashboardPage_text_noApis": "There are no APIs that you manage.",
   "dashboardPage_text_noProxyBackends": "You have not yet connected your API to a proxy. To see analytics of your API usage, browse your API from Catalog and configure a proxy from the API profile.",
   "dashboardView_text_elasticsearchError": "Something went wrong. Please try again",
+  "dateRangePicker_optionText_yesterday": "Yesterday",
+  "dateRangePicker_optionText_7days": "Last 7 Days",
+  "dateRangePicker_optionText_30days": "Last 30 Days",
   "deleteAccount_closeButton": "Close",
   "deleteAccount_confirmDelete": "Delete",
   "deleteAccount_success_message": "Account deleted",
