+++ apinf_packages/dashboard/client/analytic/analytic.html
@@ -7,38 +7,11 @@ https://joinup.ec.europa.eu/community/eupl/og_page/european-union-public-licence
   <div class="api-view">
     <div class="container">
       {{# if Template.subscriptionsReady }}
-        <h1>
-          {{ api.name }}
-        </h1>
-
-        <!-- Toolbar -->
-        {{> apiAnalyticToolbar proxyName=proxyName }}
-
-        <!-- Data is not ready -->
-        {{# unless fetchingData }}
-          {{> spinner }}
-        {{ else }}
-          <!-- If error returns -->
-          {{# if error }}
-            <h1>
-              {{_ 'apiAnalyticPage_text_elasticsearchError' }}
-            </h1>
-            <p>
-              {{ error }}
-            </p>
-          {{/ if }}
-
-          <!-- Request path has metrics data -->
-          {{# if elasticsearchData }}
-            {{# if dataAvailable }}
-              {{> apiAnalyticView elasticsearchData=elasticsearchData proxyBackendId=proxyBackendId }}
-            {{ else }}
-              <p class="alert alert-warning text-center overview-col-title">
-                {{_ 'apiAnalyticPage_warningText_noData' }}
-              </p>
-            {{/ if }}
-          {{/ if }}
-        {{/ unless }} <!-- /fetchingData -->
+        {{> apiAnalyticPageHeader proxyBackendId=proxyBackendId }}
+        <div class="form-group">
+          {{> dateRangePicker }}
+        </div>
+        {{> apiAnalyticPageBody proxyBackendId=proxyBackendId }}
       {{ else }}
         {{> spinner }}
       {{/ if }} <!-- /subscriptionReady -->
