+++ apinf_packages/dashboard/client/analytic/header/header.js
@@ -1,0 +1,23 @@
+
+/* Copyright 2017 Apinf Oy
+ This file is covered by the EUPL license.
+ You may obtain a copy of the licence at
+ https://joinup.ec.europa.eu/community/eupl/og_page/european-union-public-licence-eupl-v11 */
+
+// Meteor packages imports
+import { Template } from 'meteor/templating';
+
+// Collection imports
+import Apis from '/apinf_packages/apis/collection';
+import Proxies from '/apinf_packages/proxies/collection';
+
+Template.apiAnalyticPageHeader.helpers({
+  api () {
+    return Apis.findOne();
+  },
+  proxyName () {
+    const proxy = Proxies.findOne();
+
+    return proxy.name;
+  },
+});
