+++ apinf_packages/dashboard/client/analytic/header/header.less
@@ -1,0 +1,8 @@
+/* Copyright 2017 Apinf Oy
+This file is covered by the EUPL license.
+You may obtain a copy of the licence at
+https://joinup.ec.europa.eu/community/eupl/og_page/european-union-public-licence-eupl-v11 */
+
+.item-list {
+  display: inline-block;
+}
