+++ apinf_packages/dashboard/client/dashboard.less
@@ -93,7 +93,8 @@
   word-break: break-all;
 }
 
-.toolbar-select {
+.toolbar-select,
+#date-range-picker {
   display: inline-block;
   width: auto;
   vertical-align: middle;
