+++ apinf_packages/dashboard/client/date_range_picker/toolbar.html
@@ -1,0 +1,18 @@
+<!-- Copyright 2017 Apinf Oy
+This file is covered by the EUPL license.
+You may obtain a copy of the licence at
+https://joinup.ec.europa.eu/community/eupl/og_page/european-union-public-licence-eupl-v11 -->
+
+<template name="dateRangePicker">
+  <select name="timeframe" class="form-control" id="date-range-picker">
+    <option value="1">
+      {{_ 'dateRangePicker_optionText_yesterday' }}
+    </option>
+    <option value="7">
+      {{_ 'dateRangePicker_optionText_7days' }}
+    </option>
+    <option value="30">
+      {{_ 'dateRangePicker_optionText_30days' }}
+    </option>
+  </select>
+</template>
